+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - `AdaptiveGraphRouting.AddPlaneModifier(string name, Plane plane, double factor)`
 - `SvgSection.SaveAsSvg`, `SvgSection.SaveAsPdf`
 - `Network.TraverseLeftWithoutLeaves()`
+- `Message.FromCurve`
 - `RoutingHintLine.IsNearby`
 - `RoutingHintLine.Affects`
 
@@ -24,6 +25,8 @@
 - `EdgeInfo.HasVerticalChange` is set obsolete.
 - `AdaptiveGraphRouting.RenderElements` is no longer paint hint lines in two different colors. Instead regular edges are paint into three groups. Weights are included to additional properties of produced elements.
 - Removed rule exception from `AdaptiveGraphRouting` that prevented vertical edges turn cost being discounter.
+- `Message.FromLine` is set obsolete.
+- In `AdaptiveGridRouting`, if there are several connection points with the same cost - choose one that is closer to the trunk.
 
 ### Fixed